+++ tests/src/errors.rs
@@ -4,10 +4,15 @@ use super::create_device;
 
 #[test]
 fn invalid_pending_frame_count() {
-    let profiler = wgpu_profiler::GpuProfiler::new(wgpu_profiler::GpuProfilerSettings {
-        max_num_pending_frames: 0,
-        ..Default::default()
-    });
+    let (_, device, _queue) = create_device(wgpu::Features::TIMESTAMP_QUERY).unwrap();
+
+    let profiler = wgpu_profiler::GpuProfiler::new(
+        &device,
+        wgpu_profiler::GpuProfilerSettings {
+            max_num_pending_frames: 0,
+            ..Default::default()
+        },
+    );
     assert!(matches!(
         profiler,
         Err(wgpu_profiler::CreationError::InvalidSettings(
@@ -23,10 +28,11 @@ fn end_frame_unclosed_query() {
     )
     .unwrap();
 
-    let mut profiler = wgpu_profiler::GpuProfiler::new(GpuProfilerSettings::default()).unwrap();
+    let mut profiler =
+        wgpu_profiler::GpuProfiler::new(&device, GpuProfilerSettings::default()).unwrap();
     let unclosed_query = {
         let mut encoder = device.create_command_encoder(&wgpu::CommandEncoderDescriptor::default());
-        let query = profiler.begin_query("open query", &mut encoder, &device);
+        let query = profiler.begin_query("open query", &mut encoder);
         profiler.resolve_queries(&mut encoder);
         query
     };
@@ -52,10 +58,11 @@ fn end_frame_unresolved_query() {
     )
     .unwrap();
 
-    let mut profiler = wgpu_profiler::GpuProfiler::new(GpuProfilerSettings::default()).unwrap();
+    let mut profiler =
+        wgpu_profiler::GpuProfiler::new(&device, GpuProfilerSettings::default()).unwrap();
     {
         let mut encoder = device.create_command_encoder(&wgpu::CommandEncoderDescriptor::default());
-        let query = profiler.begin_query("open query", &mut encoder, &device);
+        let query = profiler.begin_query("open query", &mut encoder);
         profiler.end_query(&mut encoder, query);
     }
 
@@ -81,10 +88,11 @@ fn change_settings_while_query_open() {
     )
     .unwrap();
 
-    let mut profiler = wgpu_profiler::GpuProfiler::new(GpuProfilerSettings::default()).unwrap();
+    let mut profiler =
+        wgpu_profiler::GpuProfiler::new(&device, GpuProfilerSettings::default()).unwrap();
 
     let mut encoder = device.create_command_encoder(&wgpu::CommandEncoderDescriptor::default());
-    let query = profiler.begin_query("open query", &mut encoder, &device);
+    let query = profiler.begin_query("open query", &mut encoder);
 
     assert_eq!(
         profiler.change_settings(GpuProfilerSettings::default()),